refactor: 사각형 생성 옵저버패턴 개선
