+++ DrawingApp/DrawingApp/Extension/NotificaationName+.swift
@@ -1,0 +1,14 @@
+//
+//  NotificaationName+.swift
+//  DrawingApp
+//
+//  Created by 조호근 on 3/25/24.
+//
+
+import Foundation
+
+extension Notification.Name {
+    static let rectangleCreated = Notification.Name("rectangleCreated")
+    static let rectangleColorChanged = Notification.Name("rectangleColorChanged")
+    static let rectangleOpacityChanged = Notification.Name("rectangleOpacityChanged")
+}
